convert documentation layout and dependent types to typescript
