+++ znai-reactjs/package.json
@@ -3,6 +3,8 @@
   "version": "0.2.0",
   "private": true,
   "devDependencies": {
+    "@types/jest": "^25.2.2",
+    "@types/react": "^16.9.35",
     "react-component-viewer": "^0.25.0",
     "react-scripts": "3.4.1",
     "typescript": "^3.8.3"
