+++ znai-reactjs/src/App.js
@@ -36,7 +36,7 @@ import {jsxDemo} from './doc-elements/jsx/Jsx.demo'
 import {snippetsDemo, snippetsTwoSidesDemo} from './doc-elements/code-snippets/Snippet.demo'
 import {inlinedCodeDemo} from './doc-elements/code-snippets/InlinedCode.demo'
 import {documentationPreparationDemo} from './screens/documentation-preparation/DocumentationPreparation.demo'
-import {setDocMeta} from './doc-elements/docMeta'
+import {setDocMeta} from './structure/docMeta'
 import {landingDemo} from './screens/landing/Landing.demo'
 import {jupyterDemo} from './doc-elements/jupyter/Jupyter.demo'
 import {tocPanelDemo} from './structure/toc/TocPanel.demo'
