+++ znai-reactjs/src/doc-elements/Documentation.js
@@ -36,7 +36,7 @@ import PresentationRegistry from './presentation/PresentationRegistry'
 
 import AllPagesAtOnce from './AllPagesAtOnce'
 
-import {mergeDocMeta} from './docMeta'
+import {mergeDocMeta} from '../structure/docMeta'
 
 import pageContentProcessor from './pageContentProcessor.js'
 
@@ -46,9 +46,8 @@ import {pageTypesRegistry} from './page/PageTypesRegistry'
 import {injectGlobalOverridesCssLink} from './CssOverrides'
 
 import {updateGlobalDocReferences} from './references/globalDocReferences'
-import {areTocItemEquals} from '../structure/tocItem'
+import {areTocItemEquals} from '../structure/TocItem'
 
-import '../layout/DocumentationLayout.css'
 import './search/Search.css'
 
 export class Documentation extends Component {
