+++ znai-reactjs/src/doc-elements/selected-text-extensions/TextSelection.js
@@ -1,4 +1,5 @@
 /*
+ * Copyright 2020 znai maintainers
  * Copyright 2019 TWO SIGMA OPEN SOURCE, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,9 +15,9 @@
  * limitations under the License.
  */
 
-import {getDocId} from '../docMeta'
+import {getDocId} from '../../structure/docMeta'
 
-class TextSelection {
+export class TextSelection {
     constructor() {
         this.listeners = []
     }
